Start writing tests
* Fixes related to tests
* Update to Webpack 4
