+++ package.json
@@ -54,16 +54,16 @@
     "eslint-config-rapid7": "^2.11.1",
     "eslint-friendly-formatter": "3.0.0",
     "eslint-loader": "2.0.0",
-    "fast-equals": "1.2.0",
+    "fast-equals": "1.2.1",
     "fast-memoize": "2.3.0",
     "flow-babel-webpack-plugin": "1.1.1",
-    "flow-bin": "^0.66.0",
-    "html-webpack-plugin": "2.30.1",
+    "flow-bin": "^0.67.1",
+    "html-webpack-plugin": "3.0.6",
     "husky": "0.14.3",
     "in-publish": "2.0.0",
     "ink-docstrap": "1.3.2",
     "jsdoc": "3.5.5",
-    "jsdoc-babel": "0.3.0",
+    "jsdoc-babel": "0.4.0",
     "jsdom": "^11.6.2",
     "jsdom-global": "3.0.2",
     "lodash": "4.17.5",
@@ -81,8 +81,9 @@
     "rimraf": "2.6.2",
     "sinon": "^4.4.2",
     "underscore": "1.8.3",
-    "webpack": "^3.10.0",
-    "webpack-dev-server": "^2.9.7"
+    "webpack": "^4.1.1",
+    "webpack-cli": "2.0.10",
+    "webpack-dev-server": "^3.1.0"
   },
   "homepage": "https://github.com/planttheidea/moize#readme",
   "keywords": [
