+++ test-old/constants-no-promise.js
