+++ webpack.config.js
@@ -10,6 +10,8 @@ module.exports = {
 
   entry: [path.resolve(__dirname, 'src', 'index.js')],
 
+  mode: 'development',
+
   module: {
     rules: [
       {
