+++ webpack.config.minified.js
@@ -8,6 +8,8 @@ const defaultConfig = require('./webpack.config');
 module.exports = Object.assign({}, defaultConfig, {
   devtool: undefined,
 
+  mode: 'production',
+
   output: Object.assign({}, defaultConfig.output, {
     filename: 'moize.min.js'
   }),
@@ -18,28 +20,6 @@ module.exports = Object.assign({}, defaultConfig, {
       minimize: true
     }),
     new webpack.optimize.OccurrenceOrderPlugin(),
-    new webpack.optimize.ModuleConcatenationPlugin(),
-    new webpack.optimize.UglifyJsPlugin({
-      comments: false,
-      compress: {
-        booleans: true,
-        comparisons: true,
-        conditionals: true,
-        dead_code: true,
-        drop_console: true,
-        drop_debugger: true,
-        evaluate: true,
-        if_return: true,
-        join_vars: true,
-        loops: true,
-        properties: true,
-        screw_ie8: true,
-        sequences: true,
-        unused: true,
-        warnings: false
-      },
-      sourceMap: false
-    }),
     new OptimizeJsPlugin({
       sourceMap: false
     })
